Reorganised Scene and Fixed Player Movement
Set up Scene to work as a 2.5D Scene, 3D Scene with a fixed 60degree camera faced down at the flat plane.

I moved the plane to be horizontally flat so that the gravity works and created the 2.5D Scene using the above method instead.

Changed some of the variables used within the player movement code so that the player moves in conventional 8-directional movement rather than rotating and propelling itself forwards.