Added Snowed-on Ground Texture
Using the below stated video tutorial, I have worked with the textures for the plane mesh used for the ground of the scene to add texture to the floor within the game.

https://www.youtube.com/watch?v=K7a4hDRxYu8&ab_channel=ExploreGameDev

Now that All elements have textures, the visuals and theming of the game have become apparent and more of the game's features can be worked on.