Implemented Team Colours
Implemented a coloured mesh outline around snowballs created by team 1 and team 2. If a snowball is created by player 1 it has a red outline, and if it is created by player 2, it has a blue outline.

Also implemented a transparent circle around the feet of each player to indicate which is player 1 (red) and which is player 2 (blue).