Player 2 Snowball Growth and World Border Bug Fix
Amended snowball_blue to allow it to work with the snowball.gd script like the red snowballs. Also Amended the snowball script to set a starting and largest scale for the snowballs.

Player Characters were not colliding with the World Border so the Collision Layer for that Node was Changed.