Exported Game as .exe and Added Background Music
Did the final export of the game into an .exe file and a pck file that was uploaded to the itch.io page as a zip file avaliable for download.

The itch.io page will go live at this link: https://oryley.itch.io/the-snowball-game when the gameplay trailer is done and on YouTube. The gameplay trailer will also be added to the Devpost page (https://devpost.com/software/thesnowballgame) that I will update.